Fix dropped EPG programs for late channel batches
